Wiped out remaining warnings in my code

+++ src/core/monitors/admixture/AdmixtureBipartitionMonitor.cpp
@@ -299,7 +299,7 @@ void AdmixtureBipartitionMonitor::printHeader() {
     // get tree object names
     AdmixtureTree* tau = &tree->getValue();
     std::vector<AdmixtureNode*> nodes;
-    for (int i = 0; i < tree->getValue().getNumberOfNodes(); i++)
+    for (size_t i = 0; i < tree->getValue().getNumberOfNodes(); i++)
         nodes.push_back(&tau->getNode(i));
     
     for (size_t i = 0; i < nodes.size(); i++)
